RestFramework/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
mysqlclient==2.2.7
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
propcache==0.4.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tencentcloud-sdk-python==3.0.1489
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
xlrd==2.0.2
yarl==1.22.0
